Common Code for Spark 2.3 and Spark 2.4 added to one place
