+++ integration/spark/pom.xml
@@ -593,6 +593,7 @@
                 <configuration>
                   <sources>
                     <source>src/main/spark2.3</source>
+                    <source>src/main/common2.3and2.4</source>
                   </sources>
                 </configuration>
               </execution>
@@ -632,6 +633,7 @@
                 <configuration>
                   <sources>
                     <source>src/main/spark2.4</source>
+                    <source>src/main/common2.3and2.4</source>
                   </sources>
                 </configuration>
               </execution>
@@ -654,6 +656,7 @@
               <excludes>
                 <exclude>src/main/spark2.4</exclude>
                 <exclude>src/main/spark2.3</exclude>
+                <exclude>src/main/common2.3and2.4</exclude>
               </excludes>
             </configuration>
           </plugin>
