+++ integration/spark/src/main/scala/org/apache/carbondata/spark/StreamingOption.scala
@@ -19,7 +19,6 @@ package org.apache.carbondata.spark
 
 import scala.collection.mutable
 
-import org.apache.spark.sql.CarbonToSparkAdapter
 import org.apache.spark.sql.streaming.Trigger
 
 import org.apache.carbondata.common.exceptions.sql.MalformedCarbonCommandException
@@ -35,7 +34,7 @@ class StreamingOption(val userInputMap: Map[String, String]) {
     val interval = userInputMap.getOrElse(
       "interval", throw new MalformedCarbonCommandException("interval must be specified"))
     trigger match {
-      case "ProcessingTime" => CarbonToSparkAdapter.getProcessingTime(interval)
+      case "ProcessingTime" => Trigger.ProcessingTime(interval)
       case others => throw new MalformedCarbonCommandException("invalid trigger: " + trigger)
     }
   }
