+++ integration/spark/src/main/scala/org/apache/carbondata/spark/load/DataLoadProcessorStepOnSpark.scala
@@ -23,7 +23,7 @@ import com.univocity.parsers.common.TextParsingException
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.TaskContext
 import org.apache.spark.broadcast.Broadcast
-import org.apache.spark.sql.{CarbonToSparkAdapter, Row}
+import org.apache.spark.sql.{Row, SparkVersionAdapter}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.GenericInternalRow
 import org.apache.spark.util.LongAccumulator
@@ -183,7 +183,7 @@ object DataLoadProcessorStepOnSpark {
     val rowConverter = new RowConverterImpl(conf.getDataFields, conf, badRecordLogger)
     rowConverter.initialize()
 
-    CarbonToSparkAdapter.addTaskCompletionListener {
+    SparkVersionAdapter.addTaskCompletionListener {
       val hasBadRecord: Boolean = CarbonBadRecordUtil.hasBadRecord(model)
       close(conf, badRecordLogger, rowConverter)
       GlobalSortHelper.badRecordsLogger(model, partialSuccessAccum, hasBadRecord)
@@ -231,7 +231,7 @@ object DataLoadProcessorStepOnSpark {
     val rowConverter = new RowConverterImpl(conf.getDataFields, conf, badRecordLogger)
     rowConverter.initialize()
 
-    CarbonToSparkAdapter.addTaskCompletionListener {
+    SparkVersionAdapter.addTaskCompletionListener {
       val hasBadRecord: Boolean = CarbonBadRecordUtil.hasBadRecord(model)
       close(conf, badRecordLogger, rowConverter)
       GlobalSortHelper.badRecordsLogger(model, partialSuccessAccum, hasBadRecord)
