+++ integration/spark/src/main/scala/org/apache/carbondata/spark/rdd/CarbonRDD.scala
@@ -23,7 +23,7 @@ import scala.reflect.ClassTag
 import org.apache.hadoop.conf.Configuration
 import org.apache.spark.{Dependency, OneToOneDependency, Partition, TaskContext}
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{CarbonToSparkAdapter, SparkSession}
+import org.apache.spark.sql.{SparkSession, SparkVersionAdapter}
 import org.apache.spark.sql.util.SparkSQLUtil
 
 import org.apache.carbondata.core.constants.CarbonCommonConstants
@@ -72,7 +72,7 @@ abstract class CarbonRDD[T: ClassTag](
   def internalCompute(split: Partition, context: TaskContext): Iterator[T]
 
   final def compute(split: Partition, context: TaskContext): Iterator[T] = {
-    CarbonToSparkAdapter.addTaskCompletionListener(ThreadLocalSessionInfo.unsetAll())
+    SparkVersionAdapter.addTaskCompletionListener(ThreadLocalSessionInfo.unsetAll())
     carbonSessionInfo.getNonSerializableExtraInfo.put("carbonConf", getConf)
     ThreadLocalSessionInfo.setCarbonSessionInfo(carbonSessionInfo)
     TaskMetricsMap.initializeThreadLocal()
