+++ integration/spark/src/main/scala/org/apache/carbondata/spark/util/CarbonScalaUtil.scala
@@ -108,13 +108,13 @@ object CarbonScalaUtil {
           if (defaultValue) {
             timeStampFormat.format(new Date())
           } else {
-            timeStampFormat.format(CarbonToSparkAdapter.stringToTime(value))
+            timeStampFormat.format(SparkVersionAdapter.stringToTime(value))
           }
         case DateType if dateFormat != null =>
           if (defaultValue) {
             dateFormat.format(new Date())
           } else {
-            dateFormat.format(CarbonToSparkAdapter.stringToTime(value))
+            dateFormat.format(SparkVersionAdapter.stringToTime(value))
           }
         case _ =>
           val convertedValue =
@@ -162,7 +162,7 @@ object CarbonScalaUtil {
             if (defaultValue) {
               timeStampFormat.format(new Date())
             } else {
-              timeStampFormat.format(CarbonToSparkAdapter.stringToTime(value))
+              timeStampFormat.format(SparkVersionAdapter.stringToTime(value))
             }
           val convertedValue =
             DataTypeUtil
@@ -174,7 +174,7 @@ object CarbonScalaUtil {
             if (defaultValue) {
               dateFormat.format(new Date())
             } else {
-              dateFormat.format(CarbonToSparkAdapter.stringToTime(value))
+              dateFormat.format(SparkVersionAdapter.stringToTime(value))
             }
           val convertedValue =
             DataTypeUtil
@@ -236,7 +236,7 @@ object CarbonScalaUtil {
     try {
       column.getDataType match {
         case CarbonDataTypes.TIMESTAMP =>
-          CarbonToSparkAdapter.timeStampToString(value.toLong * 1000)
+          SparkVersionAdapter.timeStampToString(value.toLong * 1000)
         case CarbonDataTypes.DATE =>
           val date = DirectDictionaryKeyGeneratorFactory.getDirectDictionaryGenerator(
             column.getDataType,
@@ -245,7 +245,7 @@ object CarbonScalaUtil {
           if (date == null) {
             null
           } else {
-            CarbonToSparkAdapter.dateToString(date.toString.toInt)
+            SparkVersionAdapter.dateToString(date.toString.toInt)
           }
         case _ => value
       }
@@ -281,9 +281,9 @@ object CarbonScalaUtil {
       }
       column.getDataType match {
         case CarbonDataTypes.TIMESTAMP =>
-          CarbonToSparkAdapter.stringToTime(value).getTime.toString
+          SparkVersionAdapter.stringToTime(value).getTime.toString
         case CarbonDataTypes.DATE =>
-          CarbonToSparkAdapter.stringToTime(value).getTime.toString
+          SparkVersionAdapter.stringToTime(value).getTime.toString
         case _ => value
       }
     } catch {
