+++ integration/spark/src/main/scala/org/apache/spark/sql/carbondata/execution/datasources/CarbonFileIndex.scala
@@ -21,7 +21,7 @@ import java.util
 import scala.collection.JavaConverters._
 
 import org.apache.hadoop.fs.Path
-import org.apache.spark.sql.{CarbonToSparkAdapter, SparkSession}
+import org.apache.spark.sql.{SparkSession, SparkVersionAdapter}
 import org.apache.spark.sql.catalyst.expressions.Expression
 import org.apache.spark.sql.execution.datasources._
 import org.apache.spark.sql.types.{AtomicType, StructType}
@@ -95,7 +95,7 @@ case class CarbonFileIndex(
       val hadoopConf = sparkSession.sessionState.newHadoopConf()
       ThreadLocalSessionInfo.setConfigurationToCurrentThread(hadoopConf)
       // convert t sparks source filter
-      val filters = CarbonToSparkAdapter.translateFilter(dataFilters)
+      val filters = SparkVersionAdapter.translateFilter(dataFilters)
       val dataTypeMap = dataSchema.map(f => f.name -> f.dataType).toMap
       // convert to carbon filter expressions
       val filter: Option[CarbonExpression] = filters.filterNot{ ref =>
