+++ integration/spark/src/main/scala/org/apache/spark/sql/execution/CarbonTakeOrderedAndProjectExec.scala
@@ -19,7 +19,7 @@ package org.apache.spark.sql.execution
 
 import org.apache.spark.rdd.RDD
 import org.apache.spark.serializer.Serializer
-import org.apache.spark.sql.{CarbonTakeOrderedAndProjectExecHelper, CarbonToSparkAdapter}
+import org.apache.spark.sql.{CarbonTakeOrderedAndProjectExecHelper, SparkVersionAdapter}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.{Attribute, NamedExpression, SortOrder, UnsafeProjection}
 import org.apache.spark.sql.catalyst.expressions.codegen.LazilyGeneratedOrdering
@@ -101,7 +101,7 @@ case class CarbonTakeOrderedAndProjectExec(
     }
     // update with modified RDD (localTopK)
     val shuffled =
-      CarbonToSparkAdapter.createShuffledRowRDD(sparkContext, localTopK, child, serializer)
+      SparkVersionAdapter.createShuffledRowRDD(sparkContext, localTopK, child, serializer)
     shuffled.mapPartitions { iter =>
       val topK = org.apache.spark.util.collection.Utils.takeOrdered(iter.map(_.copy()), limit)(ord)
       if (projectList != child.output) {
