+++ integration/spark/src/main/scala/org/apache/spark/sql/execution/command/management/CommonLoadUtils.scala
@@ -1053,7 +1053,7 @@ object CommonLoadUtils {
         opt,
         loadParams.currPartitions)
       val convertedPlan =
-        CarbonToSparkAdapter.getInsertIntoCommand(
+        SparkVersionAdapter.getInsertIntoCommand(
           table = convertRelation,
           partition = loadParams.finalPartition,
           query = query,
