+++ integration/spark/src/main/scala/org/apache/spark/sql/execution/command/mutation/IUDCommonUtil.scala
@@ -44,7 +44,7 @@ object IUDCommonUtil {
       case unresolvedRelation: UnresolvedRelation =>
         val dbAndTb =
           sparkSession.sessionState.catalog.asInstanceOf[HiveSessionCatalog].getCurrentDatabase +
-          "." + CarbonToSparkAdapter.getTableIdentifier(unresolvedRelation).get.table
+          "." + SparkVersionAdapter.getTableIdentifier(unresolvedRelation).get.table
         val segmentProperties = carbonProperties
           .getProperty(CarbonCommonConstants.CARBON_INPUT_SEGMENTS + dbAndTb, "")
         if (!(segmentProperties.equals("") || segmentProperties.trim.equals("*"))) {
