+++ integration/spark/src/main/scala/org/apache/spark/sql/execution/strategy/DMLStrategy.scala
@@ -23,7 +23,7 @@ import scala.collection.mutable
 
 import org.apache.log4j.Logger
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.{CarbonCountStar, CarbonDatasourceHadoopRelation, CarbonToSparkAdapter, CountStarPlan, InsertIntoCarbonTable, SparkSession}
+import org.apache.spark.sql.{CarbonCountStar, CarbonDatasourceHadoopRelation, CarbonToSparkAdapter, CountStarPlan, InsertIntoCarbonTable, SparkSession, SparkVersionAdapter}
 import org.apache.spark.sql.catalyst.{CatalystTypeConverters, InternalRow}
 import org.apache.spark.sql.catalyst.analysis.UnresolvedAlias
 import org.apache.spark.sql.catalyst.expressions.{Alias, Ascending, Attribute, AttributeReference, Cast, Descending, Expression, IntegerLiteral, Literal, NamedExpression, ScalaUDF, SortOrder, UnsafeProjection}
@@ -133,7 +133,7 @@ object DMLStrategy extends SparkStrategy {
           val sparkSession = SparkSQLUtil.getSparkSession
           lazy val analyzer = sparkSession.sessionState.analyzer
           lazy val optimizer = sparkSession.sessionState.optimizer
-          val analyzedPlan = CarbonToSparkAdapter.invokeAnalyzerExecute(
+          val analyzedPlan = SparkVersionAdapter.invokeAnalyzerExecute(
             analyzer, rightSide)
           val polygonTablePlan = optimizer.execute(analyzedPlan)
           // transform join condition by replacing polygon column with ToRangeListAsString udf
@@ -200,7 +200,7 @@ object DMLStrategy extends SparkStrategy {
           leftKeys: Seq[Expression],
           rightKeys: Seq[Expression],
           Inner,
-          CarbonToSparkAdapter.getBuildRight,
+          SparkVersionAdapter.getBuildRight,
           carbonChild,
           planLater(right),
           condition)
@@ -215,7 +215,7 @@ object DMLStrategy extends SparkStrategy {
             leftKeys: Seq[Expression],
             rightKeys: Seq[Expression],
             Inner,
-            CarbonToSparkAdapter.getBuildLeft,
+            SparkVersionAdapter.getBuildLeft,
             planLater(left),
             carbon,
             condition)
@@ -229,7 +229,7 @@ object DMLStrategy extends SparkStrategy {
           leftKeys: Seq[Expression],
           rightKeys: Seq[Expression],
           LeftSemi,
-          CarbonToSparkAdapter.getBuildRight,
+          SparkVersionAdapter.getBuildRight,
           planLater(left),
           planLater(right),
           condition)
