+++ integration/spark/src/test/scala/org/apache/spark/carbondata/TestStreamingTableQueryFilter.scala
@@ -27,7 +27,7 @@ import scala.collection.mutable
 import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.hive.CarbonRelation
-import org.apache.spark.sql.streaming.StreamingQuery
+import org.apache.spark.sql.streaming.{StreamingQuery, Trigger}
 import org.apache.spark.sql.test.util.QueryTest
 import org.scalatest.BeforeAndAfterAll
 
@@ -184,7 +184,7 @@ class TestStreamingTableQueryFilter extends QueryTest with BeforeAndAfterAll {
           // repartition to simulate an empty partition when readSocketDF has only one row
           qry = readSocketDF.repartition(2).writeStream
             .format("carbondata")
-            .trigger(CarbonToSparkAdapter.getProcessingTime(s"$intervalSecond seconds"))
+            .trigger(Trigger.ProcessingTime(s"$intervalSecond seconds"))
             .option("checkpointLocation",
               CarbonTablePath.getStreamingCheckpointDir(carbonTable.getTablePath))
             .option("dbName", tableIdentifier.database.get)
