+++ integration/spark/src/test/scala/org/apache/spark/carbondata/TestStreamingTableWithRowParser.scala
@@ -24,10 +24,10 @@ import java.sql.{Date, Timestamp}
 
 import scala.collection.mutable
 
-import org.apache.spark.sql.{CarbonEnv, CarbonToSparkAdapter, Row, SparkSession}
+import org.apache.spark.sql.{CarbonEnv, Row, SparkSession, SparkVersionAdapter}
 import org.apache.spark.sql.catalyst.TableIdentifier
 import org.apache.spark.sql.hive.CarbonRelation
-import org.apache.spark.sql.streaming.StreamingQuery
+import org.apache.spark.sql.streaming.{StreamingQuery, Trigger}
 import org.apache.spark.sql.test.util.QueryTest
 import org.scalatest.BeforeAndAfterAll
 
@@ -797,7 +797,7 @@ class TestStreamingTableWithRowParser extends QueryTest with BeforeAndAfterAll {
           // Write data from socket stream to carbondata file
           qry = readSocketDF.writeStream
             .format("carbondata")
-            .trigger(CarbonToSparkAdapter.getProcessingTime(s"$intervalSecond seconds"))
+            .trigger(Trigger.ProcessingTime(s"$intervalSecond seconds"))
             .option("checkpointLocation", CarbonTablePath.getStreamingCheckpointDir(tablePath))
             .option("bad_records_action", badRecordAction)
             .option("dbName", tableIdentifier.database.get)
